Temp change for constant arguments to truedivide.
This implements a special case of constant arguments to
build in functions. Only tested for bert large. Totry to
mitigate any regressions tests are done for "truediv" op.

A full solution will involve inserting a placeholder tensor op in
place of the scalar, but this should work for the moment as a quick
hack.