This will fix this issue of failing jobs on merges
Currently merge cuases a glitchy failure because
the branch being merged is deleted before the downstream
mlreferences job can check out the code.

This runs the entire pipeline before success, so it
no more failures. Comes at expense of running mlreferences
and mltools in parallel, could revisit this to have a different
logic for merge pipelines.